make todos routes private

+++ server/server.js
@@ -39,7 +39,7 @@ app.post('/users/signup', (req, res) => {
   User.findOne({ email: body.email })
     .then(u => {
       if (u) {
-        return Promise.reject('Email already exists!!')
+        return Promise.reject('Email already exists!')
       } else {
         user
           .save()
@@ -85,8 +85,10 @@ app.delete('/users/logout', authenticate, (req, res) => {
 })
 
 // Todos routes
-app.get('/todos', (req, res) => {
-  Todo.find().then(
+app.get('/todos', authenticate, (req, res) => {
+  Todo.find({
+    _creator: req.user._id
+  }).then(
     todos => {
       res.send({ todos })
     },
@@ -96,14 +98,17 @@ app.get('/todos', (req, res) => {
   )
 })
 
-app.get('/todos/:id', (req, res) => {
+app.get('/todos/:id', authenticate, (req, res) => {
   var id = req.params.id
 
   if (!ObjectID.isValid(id)) {
     return res.status(404).send()
   }
 
-  Todo.findById(id)
+  Todo.findOne({
+    _id: id,
+    _creator: req.user._id
+  })
     .then(todo => {
       if (!todo) {
         return res.status(404).send()
@@ -116,9 +121,10 @@ app.get('/todos/:id', (req, res) => {
     })
 })
 
-app.post('/todos', (req, res) => {
+app.post('/todos', authenticate, (req, res) => {
   var todo = new Todo({
-    text: req.body.text
+    text: req.body.text,
+    _creator: req.user._id
   })
 
   todo.save().then(
@@ -131,9 +137,10 @@ app.post('/todos', (req, res) => {
   )
 })
 
-app.patch('/todos/:id', (req, res) => {
+app.patch('/todos/:id', authenticate, (req, res) => {
   var id = req.params.id
   var body = _.pick(req.body, ['text', 'completed'])
+
   if (!ObjectID.isValid(id)) {
     return res.status(404).send()
   }
@@ -145,7 +152,11 @@ app.patch('/todos/:id', (req, res) => {
     body.completedAt = null
   }
 
-  Todo.findByIdAndUpdate(id, { $set: body }, { new: true })
+  Todo.findOneAndUpdate(
+    { _id: id, _creator: req.user._id },
+    { $set: body },
+    { new: true }
+  )
     .then(todo => {
       if (!todo) {
         return res.status(404).send()
@@ -158,8 +169,8 @@ app.patch('/todos/:id', (req, res) => {
     })
 })
 
-app.delete('/todos', (req, res) => {
-  Todo.deleteMany({}).then(
+app.delete('/todos', authenticate, (req, res) => {
+  Todo.deleteMany({ _creator: req.user._id }).then(
     doc => {
       res.send(doc)
     },
@@ -169,18 +180,22 @@ app.delete('/todos', (req, res) => {
   )
 })
 
-app.delete('/todos/:id', (req, res) => {
+app.delete('/todos/:id', authenticate, (req, res) => {
   var id = req.params.id
 
   if (!ObjectID.isValid(id)) {
     return res.status(404).send()
   }
 
-  Todo.findByIdAndDelete(id)
+  Todo.findOneAndRemove({
+    _id: id,
+    _creator: req.user._id
+  })
     .then(todo => {
       if (!todo) {
         return res.status(404).send()
       }
+
       res.send({ todo })
     })
     .catch(e => {
